pyproject.toml
[tool.poetry]
name = "transformers-tutorial"
version = "0.1.0"
description = ""
authors = ["rym903 <nakagawa.ryoma.84n@gmail.com>"]
readme = "README.md"
packages = [{include = "transformers_tutorial"}]

[tool.poetry.dependencies]
python = "==3.8.14"
# torchの最新版をインストールするとバグる。M1 macとかの仕様らしい...?
torch = "1.11.0"
numpy = "^1.23.4"
pandas = "^1.5.1"
requests = "^2.28.1"
transformers = "^4.24.0"
datasets = "^2.7.0"
pillow = "^9.3.0"
black = "^22.10.0"
jupyter = "^1.0.0"
timm = "^0.6.11"
seaborn = "^0.12.1"
vit-pytorch = "^0.38.1"
tweepy = "^4.12.1"
python-dotenv = "^0.21.0"
lxml = "^4.9.1"
selenium = "3.13.0"
google-auth = "^2.14.1"
google-cloud-storage = "^2.6.0"
google-cloud-secret-manager = "^2.12.6"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
